refactor: replace register_analyzer() calls with @analyzer decorator
Add @analyzer(category) class decorator for cleaner analyzer registration.
Analyzers now declare their category at the class definition rather than
with a separate function call at module end.

Also fixes test_cnn_key.py imports after cnn_key.py -> musical_cnn_key.py rename.
