+++ src/providers/__init__.py
@@ -1,7 +1,7 @@
 import importlib
 import pkgutil
 from enum import Enum
-from typing import Dict, List, Type
+from typing import Callable, Dict, List, Type
 from providers.analysis import AnalyzerBase, AnalyzerCategory
 from util.debug import is_debug_mode
 from util.logging import log
@@ -108,6 +108,24 @@ def register_analyzer(category: AnalyzerCategory, klass: Type[AnalyzerBase]):
             log.debug(f"Registered {len(resources)} resources for {klass.__name__}")
     except Exception as e:
         log.warning(f"Error registering resources for {klass.__name__}: {e}")
+
+
+def analyzer(category: AnalyzerCategory) -> Callable[[Type[AnalyzerBase]], Type[AnalyzerBase]]:
+    """
+    Class decorator that registers an analyzer with the provider registry.
+
+    Usage:
+        @analyzer(AnalyzerCategory.BPM)
+        class MyBPMAnalyzer(AnalyzerBase):
+            ...
+
+    :param category: The category under which the analyzer should be registered.
+    :return: A decorator function that registers the class and returns it unchanged.
+    """
+    def decorator(klass: Type[AnalyzerBase]) -> Type[AnalyzerBase]:
+        register_analyzer(category, klass)
+        return klass
+    return decorator
 
 def discover_providers():
     """
