+++ src/providers/analysis/bpm/librosa_bpm.py
@@ -13,12 +13,13 @@ if TYPE_CHECKING:
     from PySide6.QtWidgets import QWidget
 
 from providers.analysis import AnalyzerBase, AnalyzerResult, AnalyzerCategory
-from providers import register_analyzer
+from providers import analyzer
 from providers.audio.format_descriptor import AudioFormatDescriptor
 from util.analyzer_options import AnalyzerOption, build_widget_from_option
 from util.logging import log
 
 
+@analyzer(AnalyzerCategory.BPM)
 class LibrosaBeatTrackingBPMAnalyzer(AnalyzerBase):
     """
     BPM analyzer using librosa's beat tracking algorithm.
@@ -290,7 +291,3 @@ class LibrosaBeatTrackingBPMAnalyzer(AnalyzerBase):
 
         widget.setLayout(main_layout)
         return widget
-
-
-# Register this analyzer with the BPM category
-register_analyzer(AnalyzerCategory.BPM, LibrosaBeatTrackingBPMAnalyzer)
