+++ src/providers/analysis/bpm/stub_bpm.py
@@ -11,11 +11,12 @@ from typing import Optional, List
 from PySide6.QtWidgets import QWidget, QVBoxLayout, QLabel, QSpinBox, QFormLayout
 
 from providers.analysis import AnalyzerBase, AnalyzerResult, AnalyzerCategory
-from providers import register_analyzer
+from providers import analyzer
 from util.analyzer_options import AnalyzerOption
 from util.logging import log
 
 
+@analyzer(AnalyzerCategory.BPM)
 class StubBPMAnalyzer(AnalyzerBase):
     """
     A stub analyzer for testing purposes.
@@ -89,5 +90,3 @@ class StubBPMAnalyzer(AnalyzerBase):
 
     # Note: get_settings_widget() is inherited from AnalyzerBase
     # and will auto-generate from get_options_metadata() (returns None since no options)
-
-register_analyzer(AnalyzerCategory.BPM, StubBPMAnalyzer)
